Membangun Model/requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
backoff==2.2.1
blinker==1.9.0
boto3==1.38.18
botocore==1.38.18
Bottleneck @ file:///C:/b/abs_816hr2khp1/croot/bottleneck_1731058648110/work
Brotli @ file:///C:/b/abs_c415aux9ra/croot/brotli-split_1736182803933/work
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
cloudpickle==3.1.1
colorama==0.4.6
contourpy @ file:///C:/b/abs_f2u2o_8s9g/croot/contourpy_1732540071787/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
dacite==1.6.0
dagshub==0.5.9
dagshub-annotation-converter==0.1.9
databricks-sdk==0.53.0
dataclasses-json==0.6.7
Deprecated==1.2.18
docker==7.1.0
exceptiongroup==1.3.0
fastapi==0.115.12
Flask==3.1.1
fonttools @ file:///C:/b/abs_4crkswws2h/croot/fonttools_1737040078745/work
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.40.1
gql==3.5.2
graphene==3.4.3
graphql-core==3.2.4
graphql-relay==3.2.0
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
imbalanced-learn==0.13.0
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.0
kiwisolver @ file:///C:/b/abs_faf90xet7a/croot/kiwisolver_1737040915779/work
lightgbm==4.6.0
lxml==5.4.0
Mako==1.3.10
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.0
mdurl==0.1.2
mkl-service==2.4.0
mkl_fft @ file:///C:/Users/dev-admin/mkl/mkl_fft_1730823082242/work
mkl_random @ file:///C:/Users/dev-admin/mkl/mkl_random_1730822522280/work
mlflow==2.22.0
mlflow-skinny==2.22.0
multidict==6.4.3
mypy_extensions==1.1.0
numexpr @ file:///C:/b/abs_05o8p7bfml/croot/numexpr_1730215959182/work
numpy @ file:///C:/b/abs_0123vcxhf8/croot/numpy_and_numpy_base_1725470331966/work/dist/numpy-2.0.1-cp310-cp310-win_amd64.whl#sha256=e824ec3b279f4c5207736c83ac2f2dc79b489e08501a4ea9e3fc178c45db289b
opentelemetry-api==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
packaging @ file:///C:/b/abs_3by6s2fa66/croot/packaging_1734472138782/work
pandas @ file:///C:/b/abs_e18w9ay4we/croot/pandas_1732735115740/work/dist/pandas-2.2.3-cp310-cp310-win_amd64.whl#sha256=6fc0a0523ea4ec3f23a2b790c738c857942b4c579479f8012ac8148d675b8975
pathvalidate==3.2.3
pillow @ file:///C:/b/abs_68t826txdy/croot/pillow_1744613085333/work
prometheus_client==0.22.0
propcache==0.3.1
protobuf==6.31.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing @ file:///C:/b/abs_40z8gyj9wi/croot/pyparsing_1731445739241/work
python-dateutil @ file:///C:/b/abs_3au_koqnbs/croot/python-dateutil_1716495777160/work
pytz @ file:///C:/b/abs_6ap4tsz1ox/croot/pytz_1713974360290/work
pywin32==310
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rsa==4.9.1
s3transfer==0.12.0
scikit-learn==1.6.1
scipy==1.15.3
seaborn @ file:///C:/b/abs_522_0_dy9p/croot/seaborn_1741187268016/work
semver==3.0.4
six @ file:///C:/b/abs_149wuyuo1o/croot/six_1744271521515/work
sklearn-compat==0.1.3
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
starlette==0.46.2
tenacity==9.1.2
threadpoolctl==3.6.0
treelib==1.7.1
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata @ file:///croot/python-tzdata_1746123641790/work
unicodedata2 @ file:///C:/b/abs_dfnftvxi4k/croot/unicodedata2_1736543771112/work
urllib3==2.4.0
uvicorn==0.34.2
waitress==3.0.2
Werkzeug==3.1.3
wrapt==1.17.2
xgboost==3.0.1
yarl==1.20.0
zipp==3.21.0
